fix: @jparnaudo suggested changes
*Add poiType
*Refactored 'Submit POI Page' to a component
*Single endpoint usage
