+++ src/components/Category/CategoryBanner.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { navigate } from 'gatsby-plugin-intl'
 import Paragraph from 'decentraland-gatsby/dist/components/Text/Paragraph'
 import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
-import { ProposalType } from '../../entities/Proposal/types'
+import { PoiType, ProposalType } from '../../entities/Proposal/types'
 import TokenList from 'decentraland-gatsby/dist/utils/dom/TokenList'
 
 import './CategoryBanner.css'
@@ -10,19 +10,16 @@ import './CategoryBanner.css'
 const icons = {
   [ProposalType.Catalyst]: require("../../images/icons/catalyst.svg"),
   [ProposalType.POI]: require("../../images/icons/poi.svg"),
-  [ProposalType.AddPOI]: require("../../images/icons/add-poi.svg"),
-  [ProposalType.RemovePOI]: require("../../images/icons/remove-poi.svg"),
+  [PoiType.AddPOI]: require("../../images/icons/add-poi.svg"),
+  [PoiType.RemovePOI]: require("../../images/icons/remove-poi.svg"),
   [ProposalType.BanName]: require("../../images/icons/ban-name.svg"),
   [ProposalType.Grant]: require("../../images/icons/grant.svg"),
   [ProposalType.Poll]: require("../../images/icons/poll.svg"),
 }
 
-export type CategoryBannerProps = Omit<
-  React.AnchorHTMLAttributes<HTMLAnchorElement>,
-  "children"
-> & {
+export type CategoryBannerProps = Omit<React.AnchorHTMLAttributes<HTMLAnchorElement>, "children"> & {
   active?: boolean
-  type: ProposalType
+  type: ProposalType | PoiType
 }
 
 export default React.memo(function CategoryBanner({
