+++ src/modules/locations.ts
@@ -1,4 +1,4 @@
-import { ProposalStatus, ProposalType } from "../entities/Proposal/types";
+import { PoiType, ProposalStatus, ProposalType } from "../entities/Proposal/types";
 import API from 'decentraland-gatsby/dist/utils/api/API'
 
 const GATSBY_BASE_URL = process.env.GATSBY_BASE_URL || '/'
@@ -78,7 +78,7 @@ export default {
   proposals: (options: Partial<ProposalListPage & ProposalListViewFilter & ProposalsStatusFilter & ProposalsTypeFilter & ProposalsModal> | URLSearchParams = {}) => url('/', options),
   proposal: (proposal: string, options: { new?: "true" } = {} ) => url(`/proposal/`, { id: proposal, ...options }),
   activity: (options: Partial<ProposalsStatusFilter & ProposalActivityFilter> | URLSearchParams = {}) => url(`/activity/`, options),
-  submit: (type?: ProposalType) => url(type ? `/submit/${String(type).replace('_','-')}/` : '/submit/', {}),
+  submit: (type?: ProposalType | PoiType) => url(type ? `/submit/${String(type).replace('_','-')}/` : '/submit/', {}),
   balance: (options: Partial<{ address: string }> = {}) => url(`/balance/`, options),
   welcome: () => url(`/welcome/`, {}),
 }
